Compiling on linux (glx); introducing bmp parsing in InternalBitmap replacing platform-specific bitmaps

+++ core/utils/progress_tracking.h
@@ -8,15 +8,15 @@
 namespace Beatmup {
 	class ProgressTracking {
 	private:
-		size_t progress;
+		unsigned int progress;
 	public:
 		static ProgressTracking DEVNULL;
 
 		ProgressTracking() : progress(0) {}
-		
+
 		inline void operator()() { progress++; }
 
-		inline size_t getProgress() const { return progress; }
+		inline unsigned int getProgress() const { return progress; }
 
 		inline void reset() { progress = 0; }
 	};
